+++ include/dbmodule.h
@@ -12,7 +12,7 @@ struct JOY;
 extern "C" {
 void init_dbmodule();
 void Render_tile(void* data);
-void Draw_tile(s16 x, s16 y, s16 w, s16 h, GXColor* color);
+void Draw_tile(int x, int y, int w, int h, GXColor* color);
 void Draw_line(Vec* p0, Vec* b, u32 color);
 void Draw_quad(Vec* pos, Vec* size, u32 color);
 void Draw_line3d(Vec* p0, Vec* p1, u32 color, int blend);
